+++ xerparser/schemas/account.py
@@ -1,10 +1,8 @@
 # xerparser
 # account.py
 
-from pydantic import BaseModel, Field, validator
 
-
-class ACCOUNT(BaseModel):
+class ACCOUNT:
     """
     A class to represent a cost account.
 
@@ -23,17 +21,18 @@ class ACCOUNT(BaseModel):
 
     """
 
-    uid: str = Field(alias="acct_id")
-    code: str = Field(alias="acct_short_name")
-    description: str | None
-    name: str = Field(alias="acct_name")
-
-    @validator("description", pre=True)
-    def empty_str_to_none(cls, value):
-        return (value, None)[value == "" or value == "ï»¿"]
+    def __init__(self, **kwargs) -> None:
+        self.uid: str = kwargs["acct_id"]
+        self.code: str = kwargs["acct_short_name"]
+        self.description: str = _check_description(kwargs["acct_desc"])
+        self.name: str = kwargs["acct_name"]
 
     def __eq__(self, __o: "ACCOUNT") -> bool:
         return self.name == __o.name and self.code == __o.code
 
     def __hash__(self) -> int:
         return hash((self.name, self.name))
+
+
+def _check_description(value: str) -> str:
+    return (value, "")[value == "" or value == "ï»¿"]